feat {fe api): implementation pagination

+++ internal/adapter/handler/response/default_response.go
@@ -16,8 +16,8 @@ type DefaultSuccessResponse struct {
 }
 
 type PaginationResponse struct {
-	TotalRecords int64 `json:"total_records"`
-	Page         int   `json:"page"`
-	PerPage      int   `json:"per_page"`
-	TotalPages   int   `json:"total_pages"`
+	TotalRecords int `json:"total_records"`
+	Page         int `json:"page"`
+	PerPage      int `json:"per_page"`
+	TotalPages   int `json:"total_pages"`
 }
